Updated configure_renderd.yml
Updated configure_renderd.yml with my current progress. I found that the ini_file module works great when you have a file with [sections] specified.

Note that I struggled a bit with editing the init file as it does not contain sections. I'm not sure if lineinfile is best to use or not